Split [LGTM] tests
