+++ services/lgtm/lgtm-test-helpers.js
@@ -1,0 +1,18 @@
+'use strict'
+
+const data = {
+  alerts: 0,
+  languages: [
+    { lang: 'cpp', grade: 'A+' },
+    { lang: 'javascript', grade: 'A' },
+    { lang: 'java', grade: 'B' },
+    { lang: 'python', grade: 'C' },
+    { lang: 'csharp', grade: 'D' },
+    { lang: 'other', grade: 'E' },
+    { lang: 'foo' },
+  ],
+}
+
+module.exports = {
+  data,
+}
